+++ CHANGELOG.md
@@ -27,6 +27,7 @@ HumHub Changelog
 - Fix #7966: ENV loading for Queue workers
 - Fix #7972: Content footer comment/like links separator spacings
 - Enh #7975: More details in the log message in case of error with AltchaCaptcha
+- Fix #7977: Skip adding of a default group when another is selected on the registration form
 
 1.18.0-beta.6 (January 8, 2026)
 -------------------------------